Refactored metadata form into multiple smaller components
